refactor(skills): simplify public API to createSkillsWorkspace + raw tables
Delete actions.ts—inline defineMutation calls into workspace.ts factory.
Drop standalone disk I/O, parse, serialize, and SkillsWorkspaceClient
exports. Public surface is now createSkillsWorkspace, skillsDefinition,
skillsTable, referencesTable, Skill, and Reference.

Add JSDoc on each action method (distinct from MCP description field).
Update workspace-api skill with .withActions() section and action JSDoc
conventions.
